tests/resources/requirements.txt
marshmallow==2.20.0
sphinx==2.1.2
pylama==7.7.1
pyyaml>=5.1.1
pytest==5.0.1
coverage==4.5.4

kubernetes==10.0.1

flask==1.1.1
flask-rest-api==0.16.1

python-logstash==0.4.6

prometheus_client==0.7.1

gunicorn==19.9.0
gunicorn[eventlet]==19.9.0
eventlet==0.25.0

pyspark
numpy
pandas

